pyproject.toml
[project]
name = "marketminds"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.12.0",
    "finnhub-python>=2.4.24",
    "flask>=3.1.2",
    "mcp[cli]>=1.13.1",
    "python-dotenv>=1.1.1",
    "typer>=0.16.1",
    "yfinance>=0.2.65",
]
